DataSourceModules/Electrical-Grid-Data/requirements.txt
cassandra-driver==3.7.1
numpy==1.11.1
pandas==0.18.1
#PeakUtils==1.0.3
protobuf==3.1.0.post1
protobuf-to-dict==0.1.0
python-dateutil==2.5.3
#pytz==2016.6.1
#scikit-learn==0.17.1
scipy==0.18.1
#seaborn==0.7.1
#statsmodels==0.6.1
#patsy==0.4.1
#pika
